# [connector/routing] Avoid unnecessary copy of the data