**Before:**

```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/connector/routingconnector/internal/plogutil
cpu: Apple M2 Max
BenchmarkMoveResourcesIfLogs-12    	  474394	      2522 ns/op	    3992 B/op	      74 allocs/op
BenchmarkMoveResourcesIfMetrics-12    	  201031	      6035 ns/op	    8408 B/op	     210 allocs/op
BenchmarkMoveResourcesIfTraces-12    	  221934	      5243 ns/op	    9112 B/op	     178 allocs/op
```

**After:**

```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/connector/routingconnector/internal/plogutil
cpu: Apple M2 Max
BenchmarkMoveResourcesIfLogs-12    	  577471	      1840 ns/op	    2392 B/op	      60 allocs/op
BenchmarkMoveResourcesIfMetrics-12    	  303607	      3911 ns/op	    4824 B/op	     148 allocs/op
BenchmarkMoveResourcesIfTraces-12    	  317731	      3872 ns/op	    5208 B/op	     132 allocs/op
```